Refactor RequestCleaner
