+++ Plugin/Framework/Search/Request/RequestCleaner.php
@@ -37,15 +37,8 @@
 namespace Nosto\Cmp\Plugin\Framework\Search\Request;
 
 use Magento\Framework\Search\Request\Cleaner;
-use Magento\Store\Model\StoreManagerInterface;
-use Nosto\Cmp\Helper\SearchEngine;
 use Nosto\Cmp\Logger\LoggerInterface;
-use Nosto\Cmp\Model\Filter\FilterBuilder;
-use Nosto\Cmp\Model\Service\Recommendation\StateAwareCategoryServiceInterface;
-use Nosto\Cmp\Plugin\Catalog\Block\ParameterResolverInterface;
-use Nosto\Cmp\Utils\CategoryMerchandising;
 use Nosto\Cmp\Utils\Search;
-use Nosto\Tagging\Helper\Account as NostoHelperAccount;
 
 class RequestCleaner
 {
@@ -61,68 +54,30 @@ class RequestCleaner
     const KEY_RESULTS_FROM = 'from';
     const KEY_RESULT_SIZE = 'size';
 
-    public static $nostoTmpSort = [5, 11, 401, 2023, 1];
+    /** @var GraphQlHandler */
+    private $graphqlHandler;
 
-    /**
-     * @var ParameterResolverInterface
-     */
-    private $parameterResolver;
+    /** @var WebHandler */
+    private $webHandler;
 
-    /**
-     * @var LoggerInterface
-     */
+    /** @var LoggerInterface  */
     private $logger;
 
     /**
-     * @var SearchEngine
-     */
-    private $searchEngineHelper;
-
-    /**
-     * @var StoreManagerInterface
-     */
-    private $storeManager;
-
-    /**
-     * @var NostoHelperAccount
-     */
-    private $accountHelper;
-
-    /**
-     * @var FilterBuilder
-     */
-    private $filterBuilder;
-
-    /**
-     * @var StateAwareCategoryServiceInterface
-     */
-    private $categoryService;
-
-    /**
-     * @param ParameterResolverInterface $parameterResolver
-     * @param SearchEngine $searchEngineHelper
-     * @param StoreManagerInterface $storeManager
-     * @param NostoHelperAccount $nostoHelperAccount
-     * @param FilterBuilder $filterBuilder
-     * @param StateAwareCategoryServiceInterface $categoryService
+     * RequestCleaner constructor.
+     *
+     * @param WebHandler $webHandler
+     * @param GraphQlHandler $graphQlHandler
      * @param LoggerInterface $logger
      */
     public function __construct(
-        ParameterResolverInterface $parameterResolver,
-        SearchEngine $searchEngineHelper,
-        StoreManagerInterface $storeManager,
-        NostoHelperAccount $nostoHelperAccount,
-        FilterBuilder $filterBuilder,
-        StateAwareCategoryServiceInterface $categoryService,
+        WebHandler $webHandler,
+        GraphQlHandler $graphQlHandler,
         LoggerInterface $logger
     ) {
-        $this->parameterResolver = $parameterResolver;
+        $this->webHandler = $webHandler;
+        $this->graphqlHandler = $graphQlHandler;
         $this->logger = $logger;
-        $this->searchEngineHelper = $searchEngineHelper;
-        $this->storeManager = $storeManager;
-        $this->accountHelper = $nostoHelperAccount;
-        $this->filterBuilder = $filterBuilder;
-        $this->categoryService = $categoryService;
     }
 
     /**
@@ -139,12 +94,12 @@ class RequestCleaner
             return $requestData;
         }
         try {
-            if (!(isset($requestData[self::KEY_QUERIES][self::KEY_BIND_TO_QUERY])
-                && isset($requestData[self::KEY_QUERIES][self::KEY_BIND_TO_QUERY]['queryReference']))
-            && !(isset($requestData[self::KEY_QUERIES][self::KEY_BIND_TO_GRAPHQL])
-                    && isset($requestData[self::KEY_QUERIES][self::KEY_BIND_TO_GRAPHQL]['queryReference'])
-                    && isset($requestData[self::KEY_FILTERS][self::KEY_CATEGORY_FILTER]))
-            ) {
+
+            if ($this->containsCatalogViewQueries($requestData)) {
+                $this->webHandler->handle($requestData);
+            } elseif ($this->containsGraphQlProductSearchQueries($requestData)) {
+                $this->graphqlHandler->handle($requestData);
+            } else {
                 $this->logger->debugCmp(
                     sprintf(
                         'Could not find %s from ES request data',
@@ -155,36 +110,6 @@ class RequestCleaner
                 );
                 return $requestData;
             }
-            if (isset($requestData[self::KEY_QUERIES][self::KEY_BIND_TO_GRAPHQL])) {
-                $this->categoryService->setCategoryInRegistry(
-                    $requestData[self::KEY_FILTERS][self::KEY_CATEGORY_FILTER][self::KEY_VALUE]
-                );
-            }
-            $this->logger->debugCmp(
-                sprintf(
-                    'Using %s as search engine',
-                    $this->searchEngineHelper->getCurrentEngine()
-                ),
-                $this
-            );
-            $productIds = $this->getCmpProductIds(
-                $this->parsePageNumber($requestData),
-                $this->parseLimit($requestData)
-            );
-            $this->cleanUpCmpSort($requestData);
-            if (empty($productIds)) {
-                $this->logger->debugCmp(
-                    'Nosto did not return products for the request',
-                    $this,
-                    $requestData
-                );
-                return $requestData;
-            }
-            $this->resetRequestData($requestData);
-            $this->applyCmpFilter(
-                $requestData,
-                $productIds
-            );
         } catch (\Exception $e) {
             $this->logger->debugCmp(
                 'Failed to apply CMP - see exception log(s) for details',
@@ -197,151 +122,20 @@ class RequestCleaner
         }
     }
 
-    /**
-     * Removes the Nosto sorting key as it's not indexed
-     *
-     * @param array $requestData
-     */
-    private function cleanUpCmpSort(array &$requestData)
-    {
-        unset($requestData['sort'][Search::findNostoSortingIndex($requestData)]);
+    private function containsCatalogViewQueries(array $requestData) {
+        if (isset($requestData[self::KEY_QUERIES][self::KEY_BIND_TO_QUERY])
+            && isset($requestData[self::KEY_QUERIES][self::KEY_BIND_TO_QUERY]['queryReference'])) {
+            return true;
+        }
+        return false;
     }
 
-    /**
-     * Applies given product ids to the query & filters
-     *
-     * @param array $requestData
-     * @param array $productIds
-     */
-    private function applyCmpFilter(array &$requestData, array $productIds)
-    {
-        $bindKey = '';
-        if (isset($requestData[self::KEY_QUERIES][self::KEY_BIND_TO_QUERY])) {
-            $bindKey = self::KEY_BIND_TO_QUERY;
-        } else if (isset($requestData[self::KEY_QUERIES][self::KEY_BIND_TO_GRAPHQL])) {
-            $bindKey = self::KEY_BIND_TO_GRAPHQL;
+    private function containsGraphQlProductSearchQueries(array $requestData) {
+        if (isset($requestData[self::KEY_QUERIES][self::KEY_BIND_TO_GRAPHQL])
+            && isset($requestData[self::KEY_QUERIES][self::KEY_BIND_TO_GRAPHQL]['queryReference'])
+            && isset($requestData[self::KEY_FILTERS][self::KEY_CATEGORY_FILTER])) {
+            return true;
         }
-
-        $requestData[self::KEY_QUERIES][$bindKey]['queryReference'][] = [
-            'clause' => 'must',
-            'ref' => 'nosto_cmp_id_search'
-        ];
-
-        $requestData[self::KEY_QUERIES][self::KEY_CMP] = [
-            'name' => 'nosto_cmp',
-            'filterReference' => [
-                [
-                    'clause' => 'must',
-                    'ref' => 'prod_id',
-                ],
-            ],
-            'type' => 'filteredQuery',
-        ];
-        $requestData['filters']['prod_id'] = [
-            'name' => 'prod_id',
-            'filterReference' => [
-                [
-                    'clause' => 'must',
-                    'ref' => 'prod_ids',
-                ],
-            ],
-            'type' => 'boolFilter',
-        ];
-        $requestData['filters']['prod_ids'] = [
-            'name' => 'prod_ids',
-            'field' => $this->getProductIdField(),
-            'type' => 'termFilter',
-            'value' => $productIds
-        ];
-    }
-
-    /**
-     * @param array $requestData
-     * @return int
-     */
-    private function parsePageNumber(array $requestData)
-    {
-        $from = $requestData[self::KEY_RESULTS_FROM];
-        if ($from < 1) {
-            return 0;
-        }
-        return (int) ceil($from / $this->parseLimit($requestData));
-    }
-
-    /**
-     * @param array $requestData
-     * @return int
-     */
-    private function parseLimit(array $requestData)
-    {
-        return (int) $requestData[self::KEY_RESULT_SIZE];
-    }
-
-    /**
-     * @param int $pageNum
-     * @param int $limit
-     * @return array|null
-     */
-    private function getCmpProductIds($pageNum, $limit)
-    {
-        try {
-            $res = $this->categoryService->getPersonalisationResult(
-                $pageNum,
-                $limit
-            );
-            return $res ? CategoryMerchandising::parseProductIds($res) : null;
-        } catch (\Exception $e) {
-            $this->logger->exception($e);
-            return null;
-        }
-    }
-
-    /**
-     * Removes queries & filters from the request data
-     *
-     * @param array $requestData
-     */
-    private function resetRequestData(array &$requestData)
-    {
-        $removedQueries = [];
-        foreach ($requestData[self::KEY_QUERIES] as $key => $definition) {
-            if ($key !== self::KEY_BIND_TO_QUERY && $key !== self::KEY_BIND_TO_GRAPHQL) {
-                $removedQueries[$key] = $key;
-                unset($requestData[self::KEY_QUERIES][$key]);
-            }
-        }
-        $removedRefs = [];
-        $bindKey = '';
-        if (isset($requestData[self::KEY_QUERIES][self::KEY_BIND_TO_QUERY])) {
-            $bindKey = self::KEY_BIND_TO_QUERY;
-        } else if (isset($requestData[self::KEY_QUERIES][self::KEY_BIND_TO_GRAPHQL])) {
-            $bindKey = self::KEY_BIND_TO_GRAPHQL;
-        }
-        // Also referencing definitions
-        foreach ($requestData[self::KEY_QUERIES][$bindKey]['queryReference'] as $refIndex => $ref) {
-            $refStr = $ref['ref'];
-            if (isset($removedQueries[$refStr])) {
-                $removedRefs[$refStr] = $refStr;
-                unset($requestData[self::KEY_QUERIES][$bindKey]['queryReference'][$refIndex]);
-            }
-        }
-        $requestData['filters'] = [];
-
-        // Reset also the start point since Nosto will only use product ids
-        $requestData[self::KEY_RESULTS_FROM] = 0;
-    }
-
-    /**
-     * Return the product id field
-     *
-     * @return string
-     */
-    private function getProductIdField()
-    {
-        if ($this->searchEngineHelper->isMysql()) {
-            return self::KEY_MYSQL_PRODUCT_ID;
-        } else {
-            return self::KEY_ES_PRODUCT_ID;
-        }
+        return false;
     }
 }
